+++ shenyu-e2e/shenyu-e2e-case/shenyu-e2e-case-apache-dubbo/src/test/java/org/apache/shenyu/e2e/testcase/apachedubbo/DataSynHttpTest.java
@@ -15,6 +15,8 @@
  * limitations under the License.
  */
 
+package org.apache.shenyu.e2e.testcase.apachedubbo;
+
 import com.fasterxml.jackson.core.JsonProcessingException;
 import org.apache.shenyu.e2e.client.admin.AdminClient;
 import org.apache.shenyu.e2e.client.gateway.GatewayClient;
@@ -29,7 +31,7 @@ import org.junit.jupiter.api.Test;
 import java.util.List;
 
 /**
- * Testing the correctness of Apollo data synchronization method.
+ * Testing the correctness of Http data synchronization method.
  */
 @ShenYuTest(
         mode = ShenYuEngineConfigure.Mode.DOCKER,
@@ -41,7 +43,7 @@ import java.util.List;
                         parameters = {
                                 @ShenYuTest.Parameter(key = "username", value = "admin"),
                                 @ShenYuTest.Parameter(key = "password", value = "123456"),
-                                @ShenYuTest.Parameter(key = "dataSyn", value = "zookeeper")
+                                @ShenYuTest.Parameter(key = "dataSyn", value = "admin_http")
                         }
                 ),
                 @ShenYuTest.ServiceConfigure(
@@ -50,13 +52,13 @@ import java.util.List;
                         baseUrl = "http://{hostname:localhost}:9195",
                         type = ShenYuEngineConfigure.ServiceType.SHENYU_GATEWAY,
                         parameters = {
-                                @ShenYuTest.Parameter(key = "dataSyn", value = "zookeeper")
+                                @ShenYuTest.Parameter(key = "dataSyn", value = "gateway_http")
                         }
                 )
         },
         dockerComposeFile = "classpath:./docker-compose.mysql.yml"
 )
-public class DataSynZookeeperTest {
+public class DataSynHttpTest {
 
     @Test
     void testDataSyn(final AdminClient adminClient, final GatewayClient gatewayClient) throws InterruptedException, JsonProcessingException {
@@ -69,5 +71,4 @@ public class DataSynZookeeperTest {
         Assertions.assertEquals(12, metaDataCacheList.size());
         Assertions.assertEquals(12, ruleCacheList.size());
     }
-
 }
