+++ shenyu-e2e/shenyu-e2e-case/shenyu-e2e-case-apache-dubbo/src/test/java/org/apache/shenyu/e2e/testcase/apachedubbo/DataSynNacosTest.java
@@ -15,10 +15,13 @@
  * limitations under the License.
  */
 
+package org.apache.shenyu.e2e.testcase.apachedubbo;
+
 import com.fasterxml.jackson.core.JsonProcessingException;
 import org.apache.shenyu.e2e.client.admin.AdminClient;
 import org.apache.shenyu.e2e.client.gateway.GatewayClient;
 import org.apache.shenyu.e2e.engine.annotation.ShenYuTest;
+import org.apache.shenyu.e2e.engine.annotation.ShenYuTest.Parameter;
 import org.apache.shenyu.e2e.engine.config.ShenYuEngineConfigure;
 import org.apache.shenyu.e2e.model.data.MetaData;
 import org.apache.shenyu.e2e.model.data.RuleCacheData;
@@ -29,7 +32,7 @@ import org.junit.jupiter.api.Test;
 import java.util.List;
 
 /**
- * Testing the correctness of Http data synchronization method.
+ * Testing the correctness of Nacos data synchronization method.
  */
 @ShenYuTest(
         mode = ShenYuEngineConfigure.Mode.DOCKER,
@@ -39,9 +42,9 @@ import java.util.List;
                         port = 9095,
                         baseUrl = "http://{hostname:localhost}:9095",
                         parameters = {
-                                @ShenYuTest.Parameter(key = "username", value = "admin"),
-                                @ShenYuTest.Parameter(key = "password", value = "123456"),
-                                @ShenYuTest.Parameter(key = "dataSyn", value = "admin_http")
+                                @Parameter(key = "username", value = "admin"),
+                                @Parameter(key = "password", value = "123456"),
+                                @Parameter(key = "dataSyn", value = "nacos")
                         }
                 ),
                 @ShenYuTest.ServiceConfigure(
@@ -50,18 +53,18 @@ import java.util.List;
                         baseUrl = "http://{hostname:localhost}:9195",
                         type = ShenYuEngineConfigure.ServiceType.SHENYU_GATEWAY,
                         parameters = {
-                                @ShenYuTest.Parameter(key = "dataSyn", value = "gateway_http")
+                                @Parameter(key = "dataSyn", value = "nacos")
                         }
                 )
         },
         dockerComposeFile = "classpath:./docker-compose.mysql.yml"
 )
-public class DataSynHttpTest {
+public class DataSynNacosTest {
 
     @Test
     void testDataSyn(final AdminClient adminClient, final GatewayClient gatewayClient) throws InterruptedException, JsonProcessingException {
         adminClient.login();
-        Thread.sleep(10000);
+        Thread.sleep(20000);
         List<MetaData> metaDataCacheList = gatewayClient.getMetaDataCache();
         List<SelectorCacheData> selectorCacheList = gatewayClient.getSelectorCache();
         List<RuleCacheData> ruleCacheList = gatewayClient.getRuleCache();
